requirements.txt
psutil==5.8.0
django
djangorestframework
djoser>=2.0.3
requests>=2.26.0
